dbmodels/requirements/requirements-test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-test.txt requirements-test.in
#
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
coverage==4.5.3           # via pytest-cov
factory-boy==2.12.0
fake-factory==0.7.4
faker==1.0.7              # via factory-boy
importlib-metadata==0.17  # via pluggy, pytest
more-itertools==7.0.0     # via pytest
packaging==19.0           # via pytest
pluggy==0.12.0            # via pytest
py==1.8.0                 # via pytest
pyparsing==2.4.0          # via packaging
pytest-cov==2.7.1
pytest==4.6.2
python-dateutil==2.8.0    # via fake-factory, faker
six==1.12.0               # via fake-factory, faker, packaging, pytest, python-dateutil
text-unidecode==1.2       # via faker
wcwidth==0.1.7            # via pytest
zipp==0.5.1               # via importlib-metadata
